Ajout espace client

+++ src/main/java/be/plomberie/demo/controller/CompteController.java
@@ -4,14 +4,7 @@ import java.util.List;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 
 import be.plomberie.demo.model.Compte;
 import be.plomberie.demo.service.CompteService;
@@ -41,7 +34,7 @@ public class CompteController {
     }
 
     @GetMapping("/delete/{id}")
-    public String deleteCompte(@PathVariable int id) {
+    public String deleteCompte(@PathVariable Long id) {
         compteService.deleteCompte(id);
         return "redirect:/comptes";
     }
@@ -61,7 +54,7 @@ public class CompteController {
 
     @ResponseBody
     @DeleteMapping("/api/{id}")
-    public void deleteCompteApi(@PathVariable int id) {
+    public void deleteCompteApi(@PathVariable Long id) {
         compteService.deleteCompte(id);
     }
 }
